Move DTO classes to the shared lib. Make generation DTO impl in shared lib to make it same as in other plugins
