+++ plugins/plugin-bitbucket/codenvy-plugin-bitbucket-pullrequest/pom.xml
@@ -37,6 +37,10 @@
             <artifactId>codenvy-plugin-pullrequest-ide</artifactId>
         </dependency>
         <dependency>
+            <groupId>com.codenvy.plugin</groupId>
+            <artifactId>codenvy-plugin-pullrequest-shared</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.google.guava</groupId>
             <artifactId>guava</artifactId>
         </dependency>
