+++ plugins/plugin-bitbucket/codenvy-plugin-bitbucket-pullrequest/src/main/java/com/codenvy/plugin/pullrequest/client/BitbucketContributionWorkflow.java
@@ -19,9 +19,9 @@ import com.codenvy.plugin.pullrequest.client.steps.AddReviewFactoryLinkStep;
 import com.codenvy.plugin.pullrequest.client.steps.AuthorizeCodenvyOnVCSHostStep;
 import com.codenvy.plugin.pullrequest.client.steps.CommitWorkingTreeStep;
 import com.codenvy.plugin.pullrequest.client.steps.CreateForkStep;
+import com.codenvy.plugin.pullrequest.client.steps.DefineExecutionConfiguration;
 import com.codenvy.plugin.pullrequest.client.steps.DefineWorkBranchStep;
 import com.codenvy.plugin.pullrequest.client.steps.DetectPullRequestStep;
-import com.codenvy.plugin.pullrequest.client.steps.DefineExecutionConfiguration;
 import com.codenvy.plugin.pullrequest.client.steps.DetermineUpstreamRepositoryStep;
 import com.codenvy.plugin.pullrequest.client.steps.GenerateReviewFactoryStep;
 import com.codenvy.plugin.pullrequest.client.steps.InitializeWorkflowContextStep;
