+++ plugins/plugin-bitbucket/codenvy-plugin-bitbucket-pullrequest/src/main/java/com/codenvy/plugin/pullrequest/client/BitbucketHostingService.java
@@ -14,15 +14,15 @@
  */
 package com.codenvy.plugin.pullrequest.client;
 
-import com.codenvy.plugin.pullrequest.client.dto.HostUser;
-import com.codenvy.plugin.pullrequest.client.dto.PullRequest;
-import com.codenvy.plugin.pullrequest.client.dto.Repository;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.HostingServiceTemplates;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.NoCommitsInPullRequestException;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.NoPullRequestException;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.NoUserForkException;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.ServiceUtil;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.VcsHostingService;
+import com.codenvy.plugin.pullrequest.shared.dto.HostUser;
+import com.codenvy.plugin.pullrequest.shared.dto.PullRequest;
+import com.codenvy.plugin.pullrequest.shared.dto.Repository;
 import com.google.gwt.regexp.shared.RegExp;
 import com.google.gwt.user.client.Window;
