+++ plugins/plugin-bitbucket/codenvy-plugin-bitbucket-pullrequest/src/test/java/com/codenvy/plugin/pullrequest/client/BitbucketHostingServiceTest.java
@@ -24,7 +24,6 @@ import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Listeners;
-import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
 
 /**
