+++ plugins/plugin-github/codenvy-plugin-github-pullrequest/pom.xml
@@ -33,6 +33,10 @@
             <artifactId>codenvy-plugin-pullrequest-ide</artifactId>
         </dependency>
         <dependency>
+            <groupId>com.codenvy.plugin</groupId>
+            <artifactId>codenvy-plugin-pullrequest-shared</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.google.guava</groupId>
             <artifactId>guava</artifactId>
         </dependency>
