+++ plugins/plugin-github/codenvy-plugin-github-pullrequest/src/main/java/com/codenvy/plugin/pullrequest/client/GitHubContributionWorkflow.java
@@ -28,10 +28,10 @@ import com.codenvy.plugin.pullrequest.client.steps.InitializeWorkflowContextStep
 import com.codenvy.plugin.pullrequest.client.steps.IssuePullRequestStep;
 import com.codenvy.plugin.pullrequest.client.steps.PushBranchOnForkStep;
 import com.codenvy.plugin.pullrequest.client.steps.PushBranchOnOriginStep;
-import com.codenvy.plugin.pullrequest.client.workflow.StepsChain;
 import com.codenvy.plugin.pullrequest.client.steps.UpdatePullRequestStep;
 import com.codenvy.plugin.pullrequest.client.workflow.Context;
 import com.codenvy.plugin.pullrequest.client.workflow.ContributionWorkflow;
+import com.codenvy.plugin.pullrequest.client.workflow.StepsChain;
 import com.google.common.base.Supplier;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
