+++ plugins/plugin-github/codenvy-plugin-github-pullrequest/src/main/java/com/codenvy/plugin/pullrequest/client/GitHubHostingService.java
@@ -14,9 +14,6 @@
  */
 package com.codenvy.plugin.pullrequest.client;
 
-import com.codenvy.plugin.pullrequest.client.dto.HostUser;
-import com.codenvy.plugin.pullrequest.client.dto.PullRequest;
-import com.codenvy.plugin.pullrequest.client.dto.Repository;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.HostingServiceTemplates;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.NoCommitsInPullRequestException;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.NoHistoryInCommonException;
@@ -25,6 +22,9 @@ import com.codenvy.plugin.pullrequest.client.vcs.hosting.NoUserForkException;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.PullRequestAlreadyExistsException;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.ServiceUtil;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.VcsHostingService;
+import com.codenvy.plugin.pullrequest.shared.dto.HostUser;
+import com.codenvy.plugin.pullrequest.shared.dto.PullRequest;
+import com.codenvy.plugin.pullrequest.shared.dto.Repository;
 import com.google.gwt.regexp.shared.RegExp;
 import com.google.gwt.user.client.Window;
 import com.google.gwt.user.client.rpc.AsyncCallback;
