+++ plugins/plugin-microsoft/codenvy-plugin-microsoft-vsts-pullrequest/pom.xml
@@ -38,6 +38,10 @@
             <artifactId>codenvy-plugin-pullrequest-ide</artifactId>
         </dependency>
         <dependency>
+            <groupId>com.codenvy.plugin</groupId>
+            <artifactId>codenvy-plugin-pullrequest-shared</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.google.guava</groupId>
             <artifactId>guava</artifactId>
         </dependency>
