+++ plugins/plugin-microsoft/codenvy-plugin-microsoft-vsts-pullrequest/src/main/java/com/codenvy/plugin/pullrequest/client/MicrosoftHostingService.java
@@ -14,13 +14,13 @@
  */
 package com.codenvy.plugin.pullrequest.client;
 
-import com.codenvy.plugin.pullrequest.client.dto.HostUser;
-import com.codenvy.plugin.pullrequest.client.dto.PullRequest;
-import com.codenvy.plugin.pullrequest.client.dto.Repository;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.NoPullRequestException;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.PullRequestAlreadyExistsException;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.ServiceUtil;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.VcsHostingService;
+import com.codenvy.plugin.pullrequest.shared.dto.HostUser;
+import com.codenvy.plugin.pullrequest.shared.dto.PullRequest;
+import com.codenvy.plugin.pullrequest.shared.dto.Repository;
 import com.google.gwt.regexp.shared.RegExp;
 import com.google.gwt.user.client.Window;
