+++ plugins/plugin-microsoft/codenvy-plugin-microsoft-vsts-pullrequest/src/main/java/com/codenvy/plugin/pullrequest/client/VstsStagesProvider.java
@@ -20,9 +20,9 @@ import com.codenvy.plugin.pullrequest.client.steps.CommitWorkingTreeStep;
 import com.codenvy.plugin.pullrequest.client.steps.DetectPullRequestStep;
 import com.codenvy.plugin.pullrequest.client.steps.IssuePullRequestStep;
 import com.codenvy.plugin.pullrequest.client.steps.PushBranchOnOriginStep;
-import com.codenvy.plugin.pullrequest.client.workflow.Step;
 import com.codenvy.plugin.pullrequest.client.steps.UpdatePullRequestStep;
 import com.codenvy.plugin.pullrequest.client.workflow.Context;
+import com.codenvy.plugin.pullrequest.client.workflow.Step;
 import com.google.common.collect.ImmutableSet;
 
 import javax.inject.Inject;
