+++ plugins/plugin-pullrequest/codenvy-plugin-pullrequest-ide/src/main/java/com/codenvy/plugin/pullrequest/client/steps/AddReviewFactoryLinkStep.java
@@ -14,10 +14,10 @@
  */
 package com.codenvy.plugin.pullrequest.client.steps;
 
-import com.codenvy.plugin.pullrequest.client.dto.Configuration;
 import com.codenvy.plugin.pullrequest.client.workflow.Context;
 import com.codenvy.plugin.pullrequest.client.workflow.Step;
 import com.codenvy.plugin.pullrequest.client.workflow.WorkflowExecutor;
+import com.codenvy.plugin.pullrequest.shared.dto.Configuration;
 import com.google.inject.Singleton;
 
 /**
