+++ plugins/plugin-pullrequest/codenvy-plugin-pullrequest-ide/src/main/java/com/codenvy/plugin/pullrequest/client/steps/AuthorizeCodenvyOnVCSHostStep.java
@@ -15,10 +15,10 @@
 package com.codenvy.plugin.pullrequest.client.steps;
 
 import com.codenvy.plugin.pullrequest.client.ContributeMessages;
-import com.codenvy.plugin.pullrequest.client.dto.HostUser;
 import com.codenvy.plugin.pullrequest.client.workflow.Context;
 import com.codenvy.plugin.pullrequest.client.workflow.Step;
 import com.codenvy.plugin.pullrequest.client.workflow.WorkflowExecutor;
+import com.codenvy.plugin.pullrequest.shared.dto.HostUser;
 import com.google.inject.Singleton;
 
 import org.eclipse.che.api.promises.client.Operation;
