+++ plugins/plugin-pullrequest/codenvy-plugin-pullrequest-ide/src/main/java/com/codenvy/plugin/pullrequest/client/steps/CommitWorkingTreeStep.java
@@ -16,10 +16,10 @@ package com.codenvy.plugin.pullrequest.client.steps;
 
 import com.codenvy.plugin.pullrequest.client.ContributeMessages;
 import com.codenvy.plugin.pullrequest.client.dialogs.commit.CommitPresenter;
-import com.codenvy.plugin.pullrequest.client.dto.Configuration;
 import com.codenvy.plugin.pullrequest.client.workflow.Context;
 import com.codenvy.plugin.pullrequest.client.workflow.Step;
 import com.codenvy.plugin.pullrequest.client.workflow.WorkflowExecutor;
+import com.codenvy.plugin.pullrequest.shared.dto.Configuration;
 import com.google.gwt.user.client.rpc.AsyncCallback;
 import com.google.inject.Singleton;
