+++ plugins/plugin-pullrequest/codenvy-plugin-pullrequest-ide/src/main/java/com/codenvy/plugin/pullrequest/client/steps/CreateForkStep.java
@@ -15,11 +15,11 @@
 package com.codenvy.plugin.pullrequest.client.steps;
 
 import com.codenvy.plugin.pullrequest.client.ContributeMessages;
+import com.codenvy.plugin.pullrequest.client.vcs.hosting.NoUserForkException;
 import com.codenvy.plugin.pullrequest.client.workflow.Context;
 import com.codenvy.plugin.pullrequest.client.workflow.Step;
 import com.codenvy.plugin.pullrequest.client.workflow.WorkflowExecutor;
-import com.codenvy.plugin.pullrequest.client.vcs.hosting.NoUserForkException;
-import com.codenvy.plugin.pullrequest.client.dto.Repository;
+import com.codenvy.plugin.pullrequest.shared.dto.Repository;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
