+++ plugins/plugin-pullrequest/codenvy-plugin-pullrequest-ide/src/main/java/com/codenvy/plugin/pullrequest/client/steps/DetermineUpstreamRepositoryStep.java
@@ -14,11 +14,11 @@
  */
 package com.codenvy.plugin.pullrequest.client.steps;
 
-import com.codenvy.plugin.pullrequest.client.dto.Repository;
 import com.codenvy.plugin.pullrequest.client.vcs.hosting.VcsHostingService;
 import com.codenvy.plugin.pullrequest.client.workflow.Context;
 import com.codenvy.plugin.pullrequest.client.workflow.Step;
 import com.codenvy.plugin.pullrequest.client.workflow.WorkflowExecutor;
+import com.codenvy.plugin.pullrequest.shared.dto.Repository;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
