+++ plugins/plugin-pullrequest/codenvy-plugin-pullrequest-ide/src/main/java/com/codenvy/plugin/pullrequest/client/steps/WaitForkOnRemoteStep.java
@@ -14,12 +14,11 @@
  */
 package com.codenvy.plugin.pullrequest.client.steps;
 
+import com.codenvy.plugin.pullrequest.client.vcs.hosting.VcsHostingServiceProvider;
 import com.codenvy.plugin.pullrequest.client.workflow.Context;
 import com.codenvy.plugin.pullrequest.client.workflow.Step;
 import com.codenvy.plugin.pullrequest.client.workflow.WorkflowExecutor;
-import com.codenvy.plugin.pullrequest.client.vcs.hosting.VcsHostingService;
-import com.codenvy.plugin.pullrequest.client.vcs.hosting.VcsHostingServiceProvider;
-import com.codenvy.plugin.pullrequest.client.dto.Repository;
+import com.codenvy.plugin.pullrequest.shared.dto.Repository;
 import com.google.gwt.user.client.Timer;
 import com.google.gwt.user.client.rpc.AsyncCallback;
 import com.google.inject.assistedinject.Assisted;
