+++ plugins/plugin-pullrequest/codenvy-plugin-pullrequest-shared/src/main/java/com/codenvy/plugin/pullrequest/shared/ContributionProjectTypeConstants.java
@@ -12,7 +12,7 @@
  * is strictly forbidden unless prior written permission is obtained
  * from Codenvy S.A..
  */
-package com.codenvy.plugin.pullrequest.projecttype.shared;
+package com.codenvy.plugin.pullrequest.shared;
 
 /**
  * Shared constants for the contribution project type.
