+++ plugins/plugin-pullrequest/codenvy-plugin-pullrequest-shared/src/main/java/com/codenvy/plugin/pullrequest/shared/dto/IssueComment.java
@@ -12,7 +12,7 @@
  * is strictly forbidden unless prior written permission is obtained
  * from Codenvy S.A..
  */
-package com.codenvy.plugin.pullrequest.client.dto;
+package com.codenvy.plugin.pullrequest.shared.dto;
 
 import org.eclipse.che.dto.shared.DTO;
