--- plugins/plugin-pullrequest/codenvy-plugin-pullrequest-shared/src/main/resources/com/codenvy/plugin/pullrequest/projecttype/ContributionProjectType.gwt.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-     [2012] - [2016] Codenvy, S.A.
-     All Rights Reserved.
-
-    NOTICE:  All information contained herein is, and remains
-    the property of Codenvy S.A. and its suppliers,
-    if any.  The intellectual and technical concepts contained
-    herein are proprietary to Codenvy S.A.
-    and its suppliers and may be covered by U.S. and Foreign Patents,
-    patents in process, and are protected by trade secret or copyright law.
-    Dissemination of this information or reproduction of this material
-    is strictly forbidden unless prior written permission is obtained
-    from Codenvy S.A..
-
--->
-<module>
-    <inherits name="com.google.gwt.user.User"/>
-    <inherits name="com.google.gwt.http.HTTP"/>
-    <inherits name="com.google.gwt.i18n.I18N"/>
-    <inherits name="com.google.gwt.json.JSON"/>
-
-    <source path="shared"/>
-</module>
